+++ src/rules/index.ts
@@ -17,6 +17,7 @@ import { plaintextHttpWithTokenRule } from "./plaintext-http-with-token.js";
 import { invalidEnvVarNameRule } from "./invalid-env-var-name.js";
 import { emptyEnvValueRule } from "./empty-env-value.js";
 import { cwdNotAbsoluteRule } from "./cwd-not-absolute.js";
+import { secretInArgsRule } from "./secret-in-args.js";
 
 export const BUILTIN_RULES: Rule[] = [
   structureRules,
@@ -37,4 +38,5 @@ export const BUILTIN_RULES: Rule[] = [
   invalidEnvVarNameRule,
   emptyEnvValueRule,
   cwdNotAbsoluteRule,
+  secretInArgsRule,
 ];